+++ app/src/main/kotlin/io/github/getsixtyfour/openpyn/map/MapControlTower.kt
@@ -33,9 +33,9 @@ import com.naver.android.svc.annotation.RequireViews
 import de.jupf.staticlog.Log
 import de.westnordost.countryboundaries.CountryBoundaries
 import io.github.getsixtyfour.openpyn.R
-import io.github.getsixtyfour.openpyn.utilities.LazyMarkerStorage
-import io.github.getsixtyfour.openpyn.utilities.PrintArray
-import io.github.getsixtyfour.openpyn.utilities.SubmitCallbackListener
+import io.github.getsixtyfour.openpyn.utils.LazyMarkerStorage
+import io.github.getsixtyfour.openpyn.utils.PrintArray
+import io.github.getsixtyfour.openpyn.utils.SubmitCallbackListener
 import io.github.getsixtyfour.openpyn.logException
 import io.github.sdsstudios.nvidiagpumonitor.listeners.OnCommandExecuteListener
 import io.github.sdsstudios.nvidiagpumonitor.model.Coordinate
