+++ app/src/main/kotlin/io/github/getsixtyfour/openpyn/map/Utilities.kt
@@ -36,18 +36,18 @@ import de.jupf.staticlog.Log
 import de.westnordost.countryboundaries.CountryBoundaries
 import io.github.getsixtyfour.openpyn.R
 import io.github.getsixtyfour.openpyn.R.array
-import io.github.getsixtyfour.openpyn.utilities.CITY
-import io.github.getsixtyfour.openpyn.utilities.COUNTRY
-import io.github.getsixtyfour.openpyn.utilities.FLAG
-import io.github.getsixtyfour.openpyn.utilities.IP
-import io.github.getsixtyfour.openpyn.utilities.LAT
-import io.github.getsixtyfour.openpyn.utilities.LOCATION
-import io.github.getsixtyfour.openpyn.utilities.LONG
-import io.github.getsixtyfour.openpyn.utilities.MultiSelectMapper
-import io.github.getsixtyfour.openpyn.utilities.NetworkInfo
-import io.github.getsixtyfour.openpyn.utilities.PrintArray
-import io.github.getsixtyfour.openpyn.utilities.THREAT
-import io.github.getsixtyfour.openpyn.utilities.createJson2
+import io.github.getsixtyfour.openpyn.utils.CITY
+import io.github.getsixtyfour.openpyn.utils.COUNTRY
+import io.github.getsixtyfour.openpyn.utils.FLAG
+import io.github.getsixtyfour.openpyn.utils.IP
+import io.github.getsixtyfour.openpyn.utils.LAT
+import io.github.getsixtyfour.openpyn.utils.LOCATION
+import io.github.getsixtyfour.openpyn.utils.LONG
+import io.github.getsixtyfour.openpyn.utils.MultiSelectMapper
+import io.github.getsixtyfour.openpyn.utils.NetworkInfo
+import io.github.getsixtyfour.openpyn.utils.PrintArray
+import io.github.getsixtyfour.openpyn.utils.THREAT
+import io.github.getsixtyfour.openpyn.utils.createJson2
 import io.github.getsixtyfour.openpyn.logException
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.android.Android
