+++ app/src/main/kotlin/io/github/getsixtyfour/openpyn/utils/MultiSelectMapper.kt
@@ -1,4 +1,4 @@
-package io.github.getsixtyfour.openpyn.utilities
+package io.github.getsixtyfour.openpyn.utils
 
 import com.abdeveloper.library.MultiSelectModelExtra
 import com.jayrave.moshi.pristineModels.Mapper
