+++ app/src/main/kotlin/io/github/getsixtyfour/openpyn/utils/NetworkInfo.kt
@@ -1,4 +1,4 @@
-package io.github.getsixtyfour.openpyn.utilities
+package io.github.getsixtyfour.openpyn.utils
 
 import android.app.Application
 import android.content.Context
