+++ app/src/main/kotlin/io/github/getsixtyfour/openpyn/utils/Toaster.kt
@@ -1,4 +1,4 @@
-package io.github.getsixtyfour.openpyn.utilities
+package io.github.getsixtyfour.openpyn.utils
 
 import android.content.Context
 import es.dmoral.toasty.Toasty
